+++ CHANGELOG.md
@@ -2,4 +2,13 @@
 
 ## [Unreleased]
 
+## [0.1.0] - 2025-09-01
+
 - Initial release
+
+### Features
+
+- Basic Notesaw syntax support (block, inline block, box)
+- Real-time rendering
+- Editor-to-preview scroll synchronization
+- Partial rendering for improved performance